backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.122.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
